requirements.txt
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
orjson==3.10.7
regex==2024.9.11
langid==1.1.6
nltk==3.9.1
spacy==3.7.5
python-dateutil==2.9.0.post0
openai==1.51.2
anthropic==0.39.0
mistralai==1.2.4
jsonschema==4.23.0
tqdm==4.66.5
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.3
keyring==25.5.0

# Phase 1: Observability & Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-json-logger==2.0.7

# Phase 1: Database
sqlalchemy==2.0.25
alembic==1.13.1

# Phase 2: Caching
redis==5.0.1
hiredis==2.3.2

# Phase 2: Utilities
tenacity==8.2.3
pybreaker==1.0.0
pyyaml==6.0.1

# Phase 5: Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
